Rename checks.ts -> wallet.ts + Update refs
